Rename middleware directory and update imports
