+++ backend/src/routes/music.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const musicController = require('../controllers/musicController');
-const auth = require('../middleware/authMiddleware');
+const auth = require('../middlewares/authMiddleware');
 
-const onlyMaster = require('../middleware/onlyMaster');
+const onlyMaster = require('../middlewares/onlyMaster');
 
 router.get('/', auth, musicController.getAll);
 router.post('/', auth, onlyMaster, musicController.create);
