+++ backend/src/routes/race.js
@@ -1,10 +1,10 @@
 const express = require('express');
 const router = express.Router();
 const raceController = require('../controllers/raceController');
-const auth = require('../middleware/authMiddleware');
+const auth = require('../middlewares/authMiddleware');
 
 // Тільки для майстра або адміна
-const onlyMaster = require('../middleware/onlyMaster');
+const onlyMaster = require('../middlewares/onlyMaster');
 
 router.get('/', auth, raceController.getAll);
 router.post('/', auth, onlyMaster, raceController.create);
